refactor(ui): reorganize feed/ into components/ and data/ subfolders
- components/*.tsx: one component per file
  (ContextInspectorPanel, FeedItemCard, FeedItemMenu, FeedList, FeedSkeletonItem,
   FeedTabView, FeedToggle, FeedViewToggle, ProvenanceChip, TagChip, TraceCandidateGroup)
- data/*.ts: pure helpers + render utilities
  (body-renderers, filter, helpers, inspector-query, meta, mount,
   observation-view, pagination, sanitize, summary-extract)
- Multi-component files split: toggles.ts -> 3 components; view.ts -> 2;
  inspector.ts -> ContextInspectorPanel + inspector-query helpers;
  mount.ts -> data/mount.ts + components/ProvenanceChip.tsx
- FeedViewOps type moved from view.ts to types.ts (shared across components)
